Remove final usage of fmt()
